PAXEXAM-71 Reuse  core timeout option.
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/pax/exam@15045 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ pax-exam-container-default/src/main/java/org/ops4j/pax/exam/container/def/internal/PaxRunnerTestContainer.java
@@ -300,7 +300,7 @@ class PaxRunnerTestContainer
      *
      * @return rmi lookup timeout
      */
-    private static Integer getRMITimeout( final Option... options )
+    private static long getRMITimeout( final Option... options )
     {
         final RBCLookupTimeoutOption[] timeoutOptions = filter( RBCLookupTimeoutOption.class, options );
         if( timeoutOptions.length > 0 )
